fix(ci): resolve all clang-tidy violations and enforce ASCII-only source files
- Set WarningsAsErrors to '*' in .clang-tidy (zero tolerance for violations)
- Fix 673 clang-tidy violations across 57 firmware source files:
  - readability-implicit-bool-conversion: explicit bool/int conversions
  - readability-redundant-casting: remove same-type casts
  - readability-magic-numbers: named typed enum constants
  - readability-function-size: split oversized functions into helpers
  - readability-uppercase-literal-suffix: 0xu->0xU, 1.0f->1.0F
  - readability-avoid-const-params-in-decls: remove const from prototypes
  - clang-analyzer-deadcode.DeadStores: remove unused assignments
  - readability-identifier-naming: fix remaining naming violations
- Exclude rx_hal/inc/rx72n_*_regs.h from analysis (hardware register maps)
- Disable performance-no-int-to-ptr (standard embedded uintptr_t->ptr pattern)
- Disable insecureAPI.DeprecatedOrUnsafeBufferHandling (snprintf IS safe)
- Pin clang-format-18 in CI via LLVM apt repo (matches devcontainer 18.1.8)
- Expand ASCII check to cover all text source file types with vendor exclusions
- Fix non-ASCII characters in 28 files across the repository

+++ star-rx72n-firmware/libs/rx_bus/src/rx_bus_gpio.c
@@ -366,8 +366,8 @@ static rx_err_t internal_gpio_init_callback(rx_bus_config_t* bus_config, void* u
   }
 
   /* Initialize GPIO pin */
-  const rx_err_t init_err = ctx->output ? gpio_set_output(bus_config->proto.gpio.pin)
-                                        : gpio_set_input(bus_config->proto.gpio.pin);
+  const rx_err_t init_err = (int)ctx->output ? gpio_set_output(bus_config->proto.gpio.pin)
+                                             : gpio_set_input(bus_config->proto.gpio.pin);
 
   if (init_err != k_rx_ok) {
     rx_log_error(s_tag, "GPIO HAL initialization failed");
@@ -453,8 +453,8 @@ static rx_err_t internal_gpio_write_callback(rx_bus_config_t* bus_config, void* 
   }
 
   /* Write GPIO value */
-  const rx_err_t write_err = ctx->value ? gpio_write_high(bus_config->proto.gpio.pin)
-                                        : gpio_write_low(bus_config->proto.gpio.pin);
+  const rx_err_t write_err = (int)ctx->value ? gpio_write_high(bus_config->proto.gpio.pin)
+                                             : gpio_write_low(bus_config->proto.gpio.pin);
 
   if (write_err != k_rx_ok) {
     rx_log_error(s_tag, "GPIO write failed");
